+++ lib/spior/runner.rb
@@ -30,8 +30,6 @@ module Spior
       end
       if @options.tor then
         Msg.head
-        Spior::Install::check_base
-        Spior::Copy::config_files
         if not @network
           @network = Spior::Network.new(@options.interface)
         end